Solve tse issues
